Move output classes to their own folder
