+++ examples/capture_stream.py
@@ -3,7 +3,8 @@
 import socket
 import time
 
-from picamera2.encoders import H264Encoder, FileOutput
+from picamera2.encoders import H264Encoder
+from picamera2.outputs import FileOutput
 from picamera2 import Picamera2
 
 picam2 = Picamera2()
