+++ picamera2/encoders/__init__.py
@@ -3,4 +3,3 @@ from .h264_encoder import H264Encoder
 from .jpeg_encoder import JpegEncoder
 from .mjpeg_encoder import MJPEGEncoder
 from .multi_encoder import MultiEncoder
-from .output import FileOutput, CircularOutput
