+++ picamera2/outputs/__init__.py
@@ -1,0 +1,3 @@
+from .output import Output
+from .fileoutput import FileOutput
+from .circularoutput import CircularOutput
